refactor(api-websockets): move errors and IWebsocketsIdentity, delete WebsocketService
