--- packages/api-websockets/src/features/WebsocketService/index.ts
@@ -1,1 +1,0 @@
-export { WebsocketService } from "./abstractions.js";
