+++ packages/api-websockets/src/types.ts
@@ -1,14 +1,12 @@
 import type { DbContext } from "@webiny/handler-db/types.js";
-import type { IWebsocketsContextObject } from "./context/abstractions/IWebsocketsContext.js";
 import type { ApiCoreContext } from "@webiny/api-core/types/core.js";
 import type { SecurityPermission } from "@webiny/api-core/types/security.js";
+import type { SecurityIdentity } from "@webiny/api-core/types/security.js";
 import type { GenericRecord } from "@webiny/api/types.js";
 
-export type { IWebsocketsContextObject };
+export type IWebsocketsIdentity = Pick<SecurityIdentity, "id" | "displayName" | "type">;
 
-export interface Context extends DbContext, ApiCoreContext {
-    websockets: IWebsocketsContextObject;
-}
+export interface Context extends DbContext, ApiCoreContext {}
 
 export interface WebsocketsPermission extends SecurityPermission {
     name: "websockets";
